Updates `apply_async` and background task mapping
- `apply_async` is always a synchronous call now
- Removes some old engine code used for background tasks.
